fix(monitor): publish work-item monitor events directly on main thread (fixes #1618)
Removes the double worker round-trip (2× structuredClone) for work-item
monitor events. WORK_ITEM_EVENT_MAP translation moves to ClaudeServer so
forwardWorkItemEvent publishes directly via onMonitorEvent on the main
thread, then still posts to the worker for session waiter dispatch.
Eliminates silent event loss when worker is null during crash/restart.
